[skelleton-1]: React skelleton
