+++ shapes-ui/package.json
@@ -9,7 +9,9 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "konva": "latest",
+    "react-konva": "latest"
   },
   "scripts": {
     "start": "react-scripts start",
