+++ shapes-ui/src/App.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import AppComponent from "./AppComponent/AppComponent";
+
+function App() {
+  return <AppComponent />;
+}
+
+export default App;
