+++ shapes-ui/src/AppComponent/InputValues.css
@@ -1,0 +1,3 @@
+.input-values-root {
+    display: flex;
+}
